requirements-dev.txt
-r requirements.txt
git+https://github.com/George-Ogden/mirror-rorrim; python_version >= "3.12"
git+https://github.com/George-Ogden/mypy-pytest; python_version >= "3.12"
git+https://github.com/George-Ogden/pytest-dbg; python_version >= "3.11"
inline-snapshot
pip
pre-commit
pytest-shard
pytest-snapshot
pytest-timeout
pytest-xdist
